fix: unnecessary group related api call during inital group page load
